Refactor parameter service and add DTOs for parameter creation
- Changed ID generation strategy in Parameter model from AUTO to IDENTITY.
- Refactored ParameterService to use CreateParameterDTO for saving parameters.
- Updated application properties to disable SQL logging and adjust actuator settings.
- Added spring-boot-starter-validation dependency for input validation.
- Refactored TargetParameterController and TargetParameterService to use DTOs.
- Introduced ApiResponseDTO for consistent API responses across services.
- Implemented GlobalExceptionHandler for centralized exception handling.
- Created new DTOs for maintenance tasks and products with validation annotations.
- Added GlobalExceptionHandler for maintenance service to handle exceptions uniformly.
- Created ApiResponseDTO for manual parameters service to standardize responses.

+++ inhabitants/species-service/src/main/java/it/f3rren/aquarium/species/Repository/ICoralRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import it.f3rren.aquarium.species.model.Coral;
 
-public interface ICoralRepository extends JpaRepository<Coral, Integer> {
+public interface ICoralRepository extends JpaRepository<Coral, Long> {
     
     @Query("SELECT c FROM Coral c ORDER BY c.commonName ASC")
     public List<Coral> findAllSortedByName();
